requirements.txt
africastalking==1.2.8
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-oidc-provider==0.8.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
future==1.0.0
idna==3.9
phonenumbers==8.13.45
psycopg2-binary==2.9.9
pycryptodomex==3.20.0
pyjwkest==1.4.2
requests==2.32.3
schema==0.7.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
